feat: added VolumeCommand to SlashCommands

+++ Commands.cs
@@ -82,7 +82,7 @@ namespace MusicBot
             await commandsCore.ReplayCommand(context);
         }
         [Command("volume")]
-        public async Task VolumeCommand(CommandContext context, int volume)
+        public async Task VolumeCommand(CommandContext context, long volume)
         {
             await commandsCore.VolumeCommand(context, volume);
         }
